API supports creating jogs
JSONAPI is getting more and more popular, so I've chosen to adopt it in
this project. The jog endpoint accepts distance as a decimal, and time
as seconds.

I chose created_at for the date of the jog since that field is very
common in rails applications and there was no requirement for saving
that field for another purpose.